05-data-platforms/my-taxi-pipeline/pipeline/assets/ingestion/requirements.txt
pandas==2.2.0
requests==2.31.0
pyarrow==15.0.0
python-dateutil==2.9.0

